NUMPY

+++ app.py
@@ -60,7 +60,7 @@ def sake():
   # 銘柄IDを取得
   'あなたが選んだお酒は「',text,'」です。'
   #銘柄IDを取得
-  brandId = df[df['name']==select_brands].index.values.astype(np.int)
+  brandId = df[df['name']==select_brands].index.values
   
   #フレーバーチャートを取得
   #flavor_charts_response = get_flavor_charts_response()
@@ -71,7 +71,7 @@ def sake():
     try:
       #df = pd.DataFrame(flavor_charts)
       #'ブランドＩＤ ',brandId.to_numpy()
-      brandId
+      brandId.astype(np.int)
       #df_flavorCharts[df_flavorCharts["brandId"]==brandId]
       #df = df.drop('brandId', axis=1)
       #df
